pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.15.0",
    "flask>=3.1.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
